requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
beautifulsoup4>=4.12
nba_api>=1.4
basketball-reference-scraper>=0.4
scikit-learn>=1.4
torch>=2.0
pyyaml>=6.0
lightgbm>=4.0
mapie>=0.8
optuna>=3.5
scipy>=1.11
lxml>=5.0
